#20  Integrated YOLO object detection with current system
Using only YOLO object detection is very slow. It's tried and couldn't have real time detection. Because of that, Yolo object prediction is not run continuously. If there is a motion detected, YOLO object prediction function is used and confirm the correctness of status of parking lots. How it works? If a motion detected on specified region in frame, YOLO object prediction function will run. All vehicles in the frame are detected by YOLO. Then center points of cars are calculated. After having center points of cars, detected motion is confimed by checking that if a center of car inside specified region of parking lot.